+++ docs/CHANGE_LOG.md
@@ -4,6 +4,7 @@
 
 #### **Binding Options:**
 - Added a new binding option called "views.statistics.showRangeNumbers", which states if the range counts should be shown in the lines (defaults to false).
+- BREAKING: Renamed the binding option "mapTogglesEnabled" to "colorRangeTogglesEnabled".
 
 <br>